Typo-ing over the newline character no longer consumes it.
Previously, if you typoed over the newline character, the text after it would jump up from below to be after the cursor, which was not visually pleasing.

Additionally, typo-ing in general doesn't visually overwrite two characters (oops!). It was only intended to overwrite the character the cursor was on.

Slightly changed the typing area's model-view interaction by folding the notion of "mistakes" into the "atCursor" point (the view renders the atCursor point in red if the model hasMistakes)

Added more sweet Jasmine tests to harden all this behavior.