Refactored the preferences loading/saving logic to do a more complicated dance to avoid pulling the selected checkbox values back out of the DOM.
Not that it's a major performance issue to go have jQuery fetch the values of six checkboxes, but it's the principle of the thing, I guess.

Switched from using ember 0.9.5 to a hand-built edge version of ember:
An annoyingly inconsistent API change was made in 0.9.5 that's fixed in trunk, namely if you use {{action}} helpers, the callback gets these parameters:
  0.9.5: (view, event, context)
  trunk: (event), with 'view' and 'context' stashed on the event

(see https://github.com/emberjs/ember.js/pull/544)

I wanted to write it as just 'event' and not have to change it later, so trunk it is.