Auto-indenting features during snippet typing:
When the 'return' key is pressed, the next line is now auto-indented as deeply as the previous line was.
  * If the next line is actually dedented in the snippet text, you'll have to backspace to undo the extra indentation.

When the 'backspace' key is pressed, if there is a block of space to the left of the cursor as big as a tab width, the backspace consumes the entire tab.

The sum of these two features provides a somewhat "TextMate-like" indenting experience, except since it's language-agnostic it doesn't make any sensible predictions about when you need a new block of indentation.

Added an 'init' method to the TypingText model so that all TypingText instances don't share data (oops).