Push TypingArea and CategoryPref in a more emberly direction
They both now have the proper naming and are instantiated by the
ember app whenever asked for via route.controllerFor or
controller.get('controllers.whatever').

Moved the snippet finished observer to the route controllers, though
it gets installed twice (once for the IndexRoute instance and once for
the SnippetPlayRoute) so there's some extra tricks to prevent
the snippet change actions from firing twice.

There is still a little global variable business going on in that
the TypingAreaController is stored via App.set('typingAreaController')
for the benefit of the TypingAreaView. This can probably be avoided
by more coding.