Remove the 'SnippetPlayingController' mixin
This mixin was used so that the controllers applications/index and
snippets/play had the same behavior, but ember apparently now only
invokes snippets/play anyway, even when the route is '/'. So the
extra complexity is not needed.

It's also possible that applications/index was never needed and
I just imagined the whole thing!