Show the full list of categories no matter how you visit the snippet index
Previously, if you went from the home page to the index, the categories
listed would be the distinct list of categories for all the snippets
in the "Your Scores" box. I think this is because Ember was optimistically
returning just the subset of records it had already cached when evaluating
the this.store.findAll('snippet') promise.

To get around this, the snippets index just makes another request for
categories.

Also: use the somewhat fancier <select> onchange syntax for the category select