Add `cursor: pointer` for table headers that sort the content
Change the snippets table layout so the cells are sized to fit
rather than all being a constant width.

To preserve the suggested width of 50% for the snippet text column,
this sets `width: 50%` on the snippet text header and `max-width: 0`
on the snippet text td. Per spec, `max-width: 0` is undefined
(and it sounds crazy) but in Chrome and Firefox it currently
achieves the desired result. SHRUG?

That spec:
https://www.w3.org/TR/CSS2/visudet.html#min-max-widths